requirements.txt
torch==1.7.0
opencv-python==4.4.0.44
torchvision==0.8.1
Pillow==8.0.1
mss